Connect Cloud Run with Cloud SQL/requirements.txt
flask
csvfile
mysql-connector-python